iot-anomaly-system/requirements.txt
kafka-python
requests
fastapi
uvicorn
pydantic
paho-mqtt